+++ src/Schema/Directives/HashDirective.php
@@ -8,6 +8,9 @@ use Nuwave\Lighthouse\Support\Contracts\DefinedDirective;
 
 class HashDirective extends BaseDirective implements ArgTransformerDirective, DefinedDirective
 {
+    /**
+     * @var \Illuminate\Contracts\Hashing\Hasher
+     */
     protected $hasher;
 
     public function __construct(Hasher $hasher)
@@ -19,18 +22,16 @@ class HashDirective extends BaseDirective implements ArgTransformerDirective, De
     {
         return /** @lang GraphQL */ <<<'SDL'
 """
-Run the `Hash::make` function on the argument it is defined on.
+Use Laravel hashing to transform an argument value.
+
+Useful for hashing passwords before inserting them into the database.
+This uses the default hashing driver defined in `config/hashing.php`.
 """
 directive @hash on ARGUMENT_DEFINITION | INPUT_FIELD_DEFINITION
 SDL;
     }
 
     /**
-     * Run Laravel's Hash::make on the argument.
-     *
-     * Useful for hashing passwords before inserting them into the database.
-     * Uses the hashing driver defined in config/hashing.php
-     *
      * @param  string  $argumentValue
      * @return string
      */